evproje/requirements.txt
Django==1.11
django-oauth-toolkit==1.0.0
django-cors-headers==2.0.1
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
numpy==1.13.1
oauthlib==2.0.2
pandas==0.20.3
psycopg2==2.7.1
requests==2.18.1
requests-oauthlib==0.8.0
scikit-learn==0.18.2
sklearn==0.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
